Change show filename to last broadcast
Use of the date of first broadcast was specifically to cater for Choral
Evensong which is broadcast on Wednesdays and repeated on Sunday. It
falls down when the broadcast is repeated from much earlier ago.
Switching to last broadcast should cater for more use cases, even if it
gives inconsistent results for rapid repeats.