Change method of generating time-of-day for BBC podcasts
The main problem was that the old version of the code was downloading
each BBC podcast twice: once simply to get the last-modified header and
the second time for saving. Recently, it also seemed like the last-
modified header was being changed when a more recent podcast was added
to the feed, causing duplicates to be downloaded.

While much more fake, the new method relies only on information from the
XML feed (thus is quicker), and should be more stable.